sched: Make iowait boost per task
Convert iowait boost mechanism to be per-task instead of a per-cpu
entity. The main motivation was later patches which tries to consolidate
how cpufreq updates are sent, but current iowait boost mechanism could
suffer as a consequence.

Generally, the mechanism has suffered from a number of problems already:

1. When a task migrates the boost will remain unnecessariily on prev CPU
   and the new CPU will need time to build the boost again.
2. Lack of per-cpu tracking means that on big.LITTLE systems we can't
   make an appropriate task placement on a bigger CPU as the boost value
   grows.
3. The current boost logic had a weird dependency on rate limit/decay.
   Accidental two cpufreq updates can nullifiy the boost as reported in
   [1].

Introduce a new iowait_boost field in task_struct and update it at every
enqueue_task() following the current logic that was done in schedutil.

The only noteworthy difference is that now the iowait boost is not
aggregated at cpu level at all. Meaning in the past several tasks could
make the iowait boost to grow. But this is no longer is the case.
I don't have any data to support how this was potentially needed in the
past. So leaving it out for now. It makes more sense that each task's
boost is considered individually. If such aggregation is required, it's
better to address the need once more info is available.

find_energy_efficient_cpu() now learnt how to honour the new
p->iowait_boost. It should be treated in a similar manner to
p->uclamp_min; and hence we just take the max() of the two.

All CPU iowait boost related code was removed from schedutil and intel
pstate governors. Intel pstate needed to learn how to use the new
per-task mechanism instead.

Also cpufreq.rst updated to remove mention of IO-wait boosting as it is
no longer part of the governor activity but rather a per-task property
that the governor handles as a min performance request from this task
akin to uclamp_min.

Intel pstate could do with an upgrade to learn how to handle these
request akin to schedutil.

[1] https://lore.kernel.org/lkml/739492e4-b9a3-4c55-82e6-60b02d489c5f@arm.com/

Signed-off-by: Qais Yousef <qyousef@layalina.io>